Try disabling pip's progress bar
Based on some upstream comments in pip's issue tracker, it looks like
pip's progress bar may be in some way related to this unicode path issue
on Python 2.7. While it's not entirely clear why that would be the case,
this is an easy thing to try with little to no drawbacks.